apps/runtime/python/pyproject.toml
[project]
name = "celerity-runtime-python"
version = "0.1.0"
description = "Celerity Python Runtime — hosts Python applications on the Celerity platform"
license = "Apache-2.0"
requires-python = ">=3.13"
dependencies = [
    "celerity-sdk[runtime,telemetry]>=0.2.1",
]

[dependency-groups]
dev = [
    "watchfiles>=1.0",
    "ruff>=0.8",
    "mypy>=1.13",
]

[tool.ruff]
target-version = "py313"
line-length = 100
src = ["."]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "RUF", # ruff-specific rules
]

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
exclude = ["app/"]
